GHA: build the container images as two flavors
This adds a new flavor: non-distroless (default) instead of only shipping distroless images.

This is needed for users that may need a non-stripped down Linux distributions (package managers, shell binaries, etc.), for example it is needed to be able to use the image inside a GHA workflow with the `less-advanced-securiry` image set under `container: ...` option of a GHA job (e.g., it fails to boot because `tail` command does not exist).